feat: show warnings for deprecations
